Rename akafka files to event_sub and event_pub
